+++ spec/integrations/pro/consumption/strategies/aj/dlq_ftr_lrj_mom_vp/with_non_recoverable_job_error_spec.rb
@@ -11,7 +11,7 @@ setup_karafka(allow_errors: true) do |config|
   config.kafka[:'max.poll.interval.ms'] = 10_000
   config.kafka[:'session.timeout.ms'] = 10_000
   # Use the non-Pro scheduler to achieve FIFO scheduling to stabilize this spec
-  config.internal.processing.scheduler = ::Karafka::Processing::Scheduler.new
+  config.internal.processing.scheduler_class = ::Karafka::Processing::Scheduler
 end
 
 class DlqConsumer < Karafka::BaseConsumer