+++ spec/lib/karafka/processing/scheduler_spec.rb
@@ -1,13 +1,13 @@
 # frozen_string_literal: true
 
 RSpec.describe_current do
-  subject(:scheduler) { described_class.new }
+  subject(:scheduler) { described_class.new(queue) }
+
+  let(:queue) { [] }
+  let(:jobs_array) { [] }
 
   describe '#schedule_consumption' do
-    subject(:schedule) { scheduler.schedule_consumption(queue, jobs_array) }
-
-    let(:queue) { [] }
-    let(:jobs_array) { [] }
+    subject(:schedule) { scheduler.schedule_consumption(jobs_array) }
 
     context 'when there are no messages' do
       it 'expect not to schedule anything' do
@@ -31,5 +31,13 @@ RSpec.describe_current do
         expect(queue).to eq(jobs_array)
       end
     end
+  end
+
+  describe '#manage' do
+    it { expect { scheduler.manage }.not_to raise_error }
+  end
+
+  describe '#clear' do
+    it { expect { scheduler.clear(rand.to_s) }.not_to raise_error }
   end
 end